+++ buildSrc/src/main/kotlin/io/provenance/DownloadProtosTask.kt
@@ -29,7 +29,7 @@ open class DownloadProtosTask : DefaultTask() {
 
     @Option(
         option = "provenance-version",
-        description = "Provenance release version (e.g. v0.2.1)"
+        description = "Provenance release version (e.g. v1.3.0)"
     )
     @Input
     var provenanceVersion: String? = null